improvement: add college(general) service
