+++ apps/api/general/src/app/college/app.module.ts
@@ -1,0 +1,10 @@
+import { Module } from "@nestjs/common";
+import { CollegeController } from "./app.controller";
+import { CollegeService } from "./app.service";
+
+@Module({
+  imports: [],
+  controllers: [CollegeController],
+  providers: [CollegeService],
+})
+export class CollegeModule {}
