+++ libs/api/controllers/src/general/index.ts
@@ -1,2 +1,3 @@
 export * from "./personal.controller";
 export * from "./pmb.controller";
+export * from "./college.controller";
