+++ libs/api/controllers/src/general/pmb.controller.ts
@@ -1,8 +1,6 @@
 import { Body, Controller, Delete, Get, Param, Patch, Post, Query } from "@nestjs/common";
 import { ApiOperation, ApiTags } from "@nestjs/swagger";
 import {
-  CreateDepartment,
-  CreateFaculty,
   CreateScholarship,
   CreateSelectionPath,
   createQuestion,
@@ -13,7 +11,6 @@ import { PMBService } from "@uninus/api/services";
 @Controller()
 export class PMBController {
   constructor(private readonly appService: PMBService) {}
-
   @ApiOperation({ summary: "Get Scholarship" })
   @Get("scholarship")
   async getScholarship(@Query() query: { search: string }) {
@@ -67,63 +64,6 @@ export class PMBController {
   async getRegistrationPath(@Query() query: { search: string; id: string }) {
     return await this.appService.getRegistrationPath(query);
   }
-
-  @ApiOperation({ summary: "Get Degree Program" })
-  @Get("degree-program")
-  async getDegreeProgram(@Query() query: { search: string }) {
-    return await this.appService.getDegreeProgram(query);
-  }
-
-  @ApiOperation({ summary: "Get Faculty" })
-  @Get("faculty")
-  async getFaculty(@Query() query: { search: string; degree_program_id: string }) {
-    return await this.appService.getFaculty(query);
-  }
-
-  @ApiOperation({ summary: "Create new Faculty" })
-  @Post("faculty")
-  async createFaculty(@Body() payload: CreateFaculty) {
-    return await this.appService.createFaculty(payload);
-  }
-
-  @ApiOperation({ summary: "Update faculty" })
-  @Patch("faculty/:id")
-  async updateFaculty(@Param("id") id: number, @Body() payload: CreateFaculty) {
-    return await this.appService.updateFaculty({ id, ...payload });
-  }
-
-  @Delete("faculty/:id")
-  @ApiOperation({ summary: "Delete Faculty" })
-  async deleteFaculty(@Param("id") id: number) {
-    return await this.appService.deleteFaculty({ id });
-  }
-
-  @ApiOperation({ summary: "Get Department" })
-  @Get("department")
-  async getDepartment(
-    @Query() query: { search: string; degree_program_id: string; faculty_id: string },
-  ) {
-    return await this.appService.getDepartment(query);
-  }
-
-  @ApiOperation({ summary: "Create new Department" })
-  @Post("department")
-  async createDepartment(@Body() payload: CreateDepartment) {
-    return await this.appService.createDepartment(payload);
-  }
-
-  @ApiOperation({ summary: "Update Department" })
-  @Patch("department/:id")
-  async updateDepartment(@Param("id") id: number, @Body() payload: CreateDepartment) {
-    return await this.appService.updateDepartment({ id, ...payload });
-  }
-
-  @Delete("department/:id")
-  @ApiOperation({ summary: "Delete Department" })
-  async deleteDepartment(@Param("id") id: number) {
-    return await this.appService.deleteDepartment({ id });
-  }
-
   @ApiOperation({ summary: "Get Total Registrans" })
   @Get("registrans")
   async getRegistrans(
