+++ libs/api/modules/src/general/index.ts
@@ -1,8 +1,8 @@
 import { Module } from "@nestjs/common";
 import { PrismaModule } from "../prisma";
 import { ClientsModule, Transport } from "@nestjs/microservices";
-import { PMBController, PersonalController } from "@uninus/api/controllers";
-import { PMBService, PersonalService } from "@uninus/api/services";
+import { CollegeController, PMBController, PersonalController } from "@uninus/api/controllers";
+import { CollegeService, PMBService, PersonalService } from "@uninus/api/services";
 
 @Module({
   imports: [
@@ -22,7 +22,7 @@ import { PMBService, PersonalService } from "@uninus/api/services";
       },
     ]),
   ],
-  controllers: [PersonalController, PMBController],
-  providers: [PersonalService, PMBService],
+  controllers: [PersonalController, PMBController, CollegeController],
+  providers: [PersonalService, PMBService, CollegeService],
 })
 export class GeneralModule {}
