+++ libs/api/services/src/general/college.service.ts
@@ -1,0 +1,90 @@
+import { Injectable, Inject } from "@nestjs/common";
+import { ClientProxy, RpcException } from "@nestjs/microservices";
+import { CreateDepartment, CreateFaculty } from "@uninus/api/dto";
+import { catchError, firstValueFrom, throwError } from "rxjs";
+
+@Injectable()
+export class CollegeService {
+  constructor(@Inject("GENERAL_SERVICE") private readonly client: ClientProxy) {}
+
+  async getDegreeProgram(payload: { search: string }) {
+    const response = await firstValueFrom(
+      this.client
+        .send("get_degree", payload)
+        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
+    );
+    return response;
+  }
+
+  async getFaculty(payload: { search: string; degree_program_id: string }) {
+    const response = await firstValueFrom(
+      this.client
+        .send("get_faculty", payload)
+        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
+    );
+    return response;
+  }
+
+  async createFaculty(payload: CreateFaculty) {
+    const response = await firstValueFrom(
+      this.client
+        .send("create_faculty", payload)
+        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
+    );
+    return response;
+  }
+
+  async updateFaculty(payload: CreateFaculty & { id: number }) {
+    const response = await firstValueFrom(
+      this.client
+        .send("update_faculty", payload)
+        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
+    );
+    return response;
+  }
+
+  async deleteFaculty(payload: { id: number }) {
+    const response = await firstValueFrom(
+      this.client
+        .send("delete_faculty", payload)
+        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
+    );
+    return response;
+  }
+
+  async getDepartment(payload: { search: string; degree_program_id: string; faculty_id: string }) {
+    const response = await firstValueFrom(
+      this.client
+        .send("get_department", payload)
+        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
+    );
+    return response;
+  }
+
+  async createDepartment(payload: CreateDepartment) {
+    const response = await firstValueFrom(
+      this.client
+        .send("create_department", payload)
+        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
+    );
+    return response;
+  }
+
+  async updateDepartment(payload: CreateDepartment & { id: number }) {
+    const response = await firstValueFrom(
+      this.client
+        .send("update_department", payload)
+        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
+    );
+    return response;
+  }
+
+  async deleteDepartment(payload: { id: number }) {
+    const response = await firstValueFrom(
+      this.client
+        .send("delete_department", payload)
+        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
+    );
+    return response;
+  }
+}
