+++ libs/api/services/src/general/index.ts
@@ -1,2 +1,3 @@
 export * from "./personal.service";
 export * from "./pmb.service";
+export * from "./college.service";
