+++ libs/api/services/src/general/pmb.service.ts
@@ -1,12 +1,6 @@
 import { Injectable, Inject } from "@nestjs/common";
 import { ClientProxy, RpcException } from "@nestjs/microservices";
-import {
-  CreateDepartment,
-  CreateFaculty,
-  CreateScholarship,
-  CreateSelectionPath,
-  createQuestion,
-} from "@uninus/api/dto";
+import { CreateScholarship, CreateSelectionPath, createQuestion } from "@uninus/api/dto";
 import { catchError, firstValueFrom, throwError } from "rxjs";
 
 @Injectable()
@@ -89,87 +83,6 @@ export class PMBService {
     const response = await firstValueFrom(
       this.client
         .send("get_registration_path", payload)
-        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
-    );
-    return response;
-  }
-
-  async getDegreeProgram(payload: { search: string }) {
-    const response = await firstValueFrom(
-      this.client
-        .send("get_degree", payload)
-        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
-    );
-    return response;
-  }
-
-  async getFaculty(payload: { search: string; degree_program_id: string }) {
-    const response = await firstValueFrom(
-      this.client
-        .send("get_faculty", payload)
-        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
-    );
-    return response;
-  }
-
-  async createFaculty(payload: CreateFaculty) {
-    const response = await firstValueFrom(
-      this.client
-        .send("create_faculty", payload)
-        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
-    );
-    return response;
-  }
-
-  async updateFaculty(payload: CreateFaculty & { id: number }) {
-    const response = await firstValueFrom(
-      this.client
-        .send("update_faculty", payload)
-        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
-    );
-    return response;
-  }
-
-  async deleteFaculty(payload: { id: number }) {
-    const response = await firstValueFrom(
-      this.client
-        .send("delete_faculty", payload)
-        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
-    );
-    return response;
-  }
-
-  async getDepartment(payload: { search: string; degree_program_id: string; faculty_id: string }) {
-    const response = await firstValueFrom(
-      this.client
-        .send("get_department", payload)
-        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
-    );
-    return response;
-  }
-
-  async createDepartment(payload: CreateDepartment) {
-    const response = await firstValueFrom(
-      this.client
-        .send("create_department", payload)
-        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
-    );
-    return response;
-  }
-
-  async updateDepartment(payload: CreateDepartment & { id: number }) {
-    const response = await firstValueFrom(
-      this.client
-        .send("update_department", payload)
-        .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
-    );
-    return response;
-  }
-
-  async deleteDepartment(payload: { id: number }) {
-    const response = await firstValueFrom(
-      this.client
-        .send("delete_department", payload)
         .pipe(catchError((error) => throwError(() => new RpcException(error.response)))),
     );
     return response;
